Change DB Size, export logs
When running on an azure container instance, if the machine restarts or
terminates on failure, ACI drops the logs. The big change here is the
inclusion of the Log Analytics Workspace credentials (ID and Key), which
allow the container to stream those logs to an external store for
posthumous inspection.

Additionally, upon inspection of said logs, discovered an interaction
between FDSLoader application and Azure
PostgreSQL flexible server's "autogrow" functionality.

Postgres Database was running out of space, even though the PostgreSQL
database is set to automatically expand storage, because autogrow
increases the space limit when there's less than 10GB available, but
that assumes you're not trying to INSERT or UPDATE a table and do more
than 10Gb of operations all in one go (which it seems the Loader does).

Bumping to 512 resolved issue, and process ran smoothly.